+++ .gitignore
@@ -4,3 +4,6 @@ excel_sheets/*
 .DS_Store
 excel_venv2/
 *.pyc
+*.egg-info/
+dist/
+build/
