+++ setup.cfg
@@ -1,0 +1,95 @@
+[metadata]
+name = invoice_automation
+version = 0.0.1
+author = Ronald Davis
+author_email = rdavis@blockchain.com
+maintainer = Ronald Davis
+maintainer_email = ronalddavisr3@gmail.com
+description = Just used for extracting the sheets from a spreadsheet and placing them in their own appropriately titled PDF
+long_description  = file: README.md
+long_description_content_type = text/markdown
+license = MIT License
+requires-python = >=3.8
+classifiers = 
+    Programming Language :: Python :: 3,
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+project_urls = 
+    Homepage = https://github.com/rdavis-bc/invoice_automation
+    Bug Tracker = https://github.com/rdavis-bc/invoice_automation/issues
+
+
+
+[options]
+package_dir =
+    =src
+zip_safe = True
+include_package_data = True
+install_requires = 
+    appnope==0.1.2
+    aspose-cells==22.3.0
+    astroid==2.11.3
+    asttokens==2.0.5
+    attrs==21.4.0
+    backcall==0.2.0
+    black==22.3.0
+    build==0.8.0
+    click==8.1.3
+    decorator==5.1.1
+    dill==0.3.4
+    et-xmlfile==1.1.0
+    executing==0.8.3
+    iniconfig==1.1.1
+    ipython==8.1.1
+    isort==5.10.1
+    jedi==0.18.1
+    JPype1==1.3.0
+    lazy-object-proxy==1.7.1
+    lxml==4.8.0
+    matplotlib-inline==0.1.3
+    mccabe==0.7.0
+    mypy==0.941
+    mypy-extensions==0.4.3
+    numpy==1.22.3
+    openpyxl==3.0.9
+    packaging==21.3
+    pandas==1.4.1
+    parso==0.8.3
+    pathspec==0.9.0
+    pdfkit==1.0.0
+    pep517==0.12.0
+    pexpect==4.8.0
+    pickleshare==0.7.5
+    pkg_resources==0.0.0
+    platformdirs==2.5.2
+    pluggy==1.0.0
+    prompt-toolkit==3.0.28
+    psutil==5.9.0
+    ptyprocess==0.7.0
+    pure-eval==0.2.2
+    py==1.11.0
+    Pygments==2.11.2
+    pylint==2.13.7
+    pyparsing==3.0.9
+    PyPDF2==1.26.0
+    pytest==7.1.2
+    python-dateutil==2.8.2
+    pytz==2021.3
+    six==1.16.0
+    stack-data==0.2.0
+    tomli==2.0.1
+    traitlets==5.1.1
+    typing_extensions==4.1.1
+    wcwidth==0.2.5
+    wrapt==1.14.0
+    xlwings==0.27.2
+
+[options.package_data]
+example = ../images/*.png, ../excel_sheets/*.xlsx
+
+[options.packages.find]
+where=src
+
+[options.entry_points]
+console_scripts = 
+    create-invoices = scripts.main:main
