--- src/logging_file.txt
@@ -1,4 +1,0 @@
-serialization taking place at 2022-04-30 19:00:03.660507
-serialization taking place at 2022-05-01 22:38:50.209987
-serialization taking place at 2022-05-01 22:41:27.268162
-serialization taking place at 2022-05-02 00:26:50.271041
